Fix compiler warnings

+++ test/base/CharStringTest.c
@@ -102,10 +102,10 @@ static int _testAppendCharStrings(void) {
 static int _testAppendCharStringsOverCapacity(void) {
   CharString a = newCharStringWithCString("a");
   CharString b = newCharStringWithCString("1234567890");
-  assertSizeEquals(a->capacity, 2l);
+  assertSizeEquals(a->capacity, (size_t)2);
   charStringAppend(a, b);
   assertCharStringEquals(a, "a1234567890");
-  assertSizeEquals(a->capacity, 12l);
+  assertSizeEquals(a->capacity, (size_t)12);
   freeCharString(a);
   freeCharString(b);
   return 0;
